requirements.txt
setuptools
python-dotenv
wheel
requests
notion-client